Got the fitter to run, but for some reason the intepolator Evaluate method is only giving 0 within the function.

+++ include/DataReading.h
@@ -5,6 +5,6 @@
 
 WCData ReadWCDataFile(const std::string&);
 
-std::vector<float> readIdealWFs(unsigned int, int, const std::string&, unsigned int);
+std::vector<double> readIdealWFs(unsigned int, int, const std::string&, unsigned int);
 
 #endif //RECOMORE_DATAREADING_H
